refactor: replace utils to lib folder
