+++ src/lib/utils/variableReplacer.ts
